+++ packages/vital-navigation/src/components/Separator/tokens/vitalSeparator.ts
@@ -35,8 +35,7 @@ const FooterMenu = asElementToken({
   Spacing: {
     _: as(
       'mb-7 pb-8 last:mb-0 last:pb-0',
-      'md:mb-0 md:pb-0 md:px-10 md:first:pl-0 md:last:pr-0',
-      'lg:px-12',
+      'md:mb-0 md:pb-0',
     ),
   },
 });